+++ sulong/projects/com.oracle.truffle.llvm.parser/src/com/oracle/truffle/llvm/parser/listeners/Module.java
@@ -186,7 +186,12 @@ public final class Module implements ParserListener {
             visibility = buffer.read();
         }
 
-        GlobalVariable global = GlobalVariable.create(isConstant, (PointerType) type, align, sectionName, linkage, visibility, scope.getSymbols(), initialiser, index.getAndIncrement());
+        long threadLocal = 0;
+        if (buffer.remaining() > 0) {
+            threadLocal = buffer.read();
+        }
+
+        GlobalVariable global = GlobalVariable.create(isConstant, (PointerType) type, align, sectionName, linkage, visibility, threadLocal, scope.getSymbols(), initialiser, index.getAndIncrement());
         assignNameFromStrTab(name, global);
         module.addGlobalVariable(global);
         scope.addSymbol(global, global.getType());