+++ sulong/projects/com.oracle.truffle.llvm.parser/src/com/oracle/truffle/llvm/parser/model/symbols/globals/GlobalVariable.java
@@ -36,8 +36,8 @@ import com.oracle.truffle.llvm.parser.model.enums.Visibility;
 import com.oracle.truffle.llvm.parser.model.visitors.SymbolVisitor;
 import com.oracle.truffle.llvm.runtime.CommonNodeFactory;
 import com.oracle.truffle.llvm.runtime.GetStackSpaceFactory;
+import com.oracle.truffle.llvm.runtime.LLVMSymbol;
 import com.oracle.truffle.llvm.runtime.datalayout.DataLayout;
-import com.oracle.truffle.llvm.runtime.global.LLVMGlobal;
 import com.oracle.truffle.llvm.runtime.nodes.api.LLVMExpressionNode;
 import com.oracle.truffle.llvm.runtime.types.PointerType;
 
@@ -49,10 +49,14 @@ public final class GlobalVariable extends GlobalValueSymbol {
 
     private final String sectionName;
 
-    private GlobalVariable(boolean isReadOnly, PointerType type, int align, String sectionName, Linkage linkage, Visibility visibility, SymbolTable symbolTable, int value, int index) {
+    private final boolean isThreadLocal;
+
+    private GlobalVariable(boolean isReadOnly, PointerType type, int align, String sectionName, Linkage linkage, Visibility visibility, boolean threadLocal, SymbolTable symbolTable, int value,
+                    int index) {
         super(type, linkage, visibility, symbolTable, value, index);
         this.isReadOnly = isReadOnly;
         this.align = align;
+        this.isThreadLocal = threadLocal;
         this.sectionName = sectionName;
     }
 
@@ -73,13 +77,25 @@ public final class GlobalVariable extends GlobalValueSymbol {
         return sectionName;
     }
 
-    public static GlobalVariable create(boolean isReadOnly, PointerType type, int align, String sectionName, long linkage, long visibility, SymbolTable symbolTable, int value, int index) {
-        return new GlobalVariable(isReadOnly, type, align, sectionName, Linkage.decode(linkage), Visibility.decode(visibility), symbolTable, value, index);
+    public static GlobalVariable create(boolean isReadOnly, PointerType type, int align, String sectionName, long linkage, long visibility, long threadLocal, SymbolTable symbolTable, int value,
+                    int index) {
+        return new GlobalVariable(isReadOnly, type, align, sectionName, Linkage.decode(linkage), Visibility.decode(visibility), threadLocal > 0, symbolTable, value, index);
+    }
+
+    public boolean isThreadLocal() {
+        return isThreadLocal;
     }
 
     @Override
     public LLVMExpressionNode createNode(LLVMParserRuntime runtime, DataLayout dataLayout, GetStackSpaceFactory stackFactory) {
-        LLVMGlobal value = runtime.lookupGlobal(getName());
-        return CommonNodeFactory.createLiteral(value, new PointerType(getType()));
+        LLVMSymbol symbol = runtime.lookupSymbol(getName());
+        if (symbol.isGlobalVariable()) {
+            symbol = symbol.asGlobalVariable();
+        } else if (symbol.isThreadLocalSymbol()) {
+            symbol = symbol.asThreadLocalSymbol();
+        } else {
+            throw new AssertionError(symbol.getClass());
+        }
+        return CommonNodeFactory.createLiteral(symbol, new PointerType(getType()));
     }
 }