+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/LLVMAlias.java
@@ -96,4 +96,14 @@ public class LLVMAlias extends LLVMSymbol {
     public LLVMElemPtrSymbol asElemPtrExpression() {
         return target.asElemPtrExpression();
     }
+
+    @Override
+    public LLVMThreadLocalSymbol asThreadLocalSymbol() {
+        return target.asThreadLocalSymbol();
+    }
+
+    @Override
+    public boolean isThreadLocalSymbol() {
+        return target.isThreadLocalSymbol();
+    }
 }